requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
email-validator
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-multipart
pandas
scikit-learn
